feat(cor): ocr
